fix(gql): normalize id() to Int, rename procedure columns to snake_case, raise APSP limit
- Change id() function to return GqlValue::Int instead of GqlValue::UInt
  so WHERE id(n) = $param comparisons work with client-supplied Int parameters
- Change node.id and edge.id property access to return Int consistently
- Rename all camelCase YIELD column names to snake_case across 12 procedure
  files: nodeId->node_id, communityId->community_id, componentId->component_id,
  nodeCount->node_count, edgeCount->edge_count, memoryType->memory_type,
  createdAt->created_at, sampleCount->sample_count, communitySize->community_size,
  communityMembers->community_members, communityLabels->community_labels
- Update all MCP tool files to match renamed YIELD columns and use as_i64()
  instead of as_u64() for node ID extraction from JSON results
- Raise APSP node limit from 1000 to 5000
- Add integration test for edge creation between nodes matched by id()

+++ crates/selene-gql/src/runtime/eval.rs
@@ -401,7 +401,7 @@ pub(crate) fn resolve_property(
         GqlValue::Node(node_id) => {
             // Special properties
             if key.as_str() == "id" {
-                return Ok(GqlValue::UInt(node_id.0));
+                return Ok(GqlValue::Int(node_id.0 as i64));
             }
             // Look up from graph
             match graph.get_node(*node_id) {
@@ -422,7 +422,7 @@ pub(crate) fn resolve_property(
         GqlValue::Edge(edge_id) => {
             // Special properties
             if key.as_str() == "id" {
-                return Ok(GqlValue::UInt(edge_id.0));
+                return Ok(GqlValue::Int(edge_id.0 as i64));
             }
             match graph.get_edge(*edge_id) {
                 Some(edge) => match edge.properties.get(key) {
@@ -839,7 +839,7 @@ mod tests {
         let g = test_graph();
         let b = binding_with_node("s", 1);
         let expr = Expr::Property(Box::new(Expr::Var(IStr::new("s"))), IStr::new("id"));
-        assert_eq!(eval_expr(&expr, &b, &g).unwrap(), GqlValue::UInt(1));
+        assert_eq!(eval_expr(&expr, &b, &g).unwrap(), GqlValue::Int(1));
     }
 
     #[test]
